pywhatkit_alt/requirements.txt
selenium
yt_dlp
pillow
opencv-python
pyautogui
pyttsx3
speechrecognition
fpdf
webdriver_manager